+++ src/printer_model_scaml.ml
@@ -187,6 +187,8 @@ let pp_model fmt (model : model) =
 
     | Clear _ -> Format.fprintf fmt "// TODO api storage: clear"
 
+    | Update _ -> Format.fprintf fmt "// TODO api storage: update"
+
     | UpdateAdd (an, fn) ->
       let k, _t = Utils.get_asset_key model an in
       let ft, c = Utils.get_field_container model an fn in
@@ -420,6 +422,8 @@ let pp_model fmt (model : model) =
     | Bisnone t -> Format.fprintf fmt "isnone on %a" pp_type t
     | Bissome t -> Format.fprintf fmt "issome on %a" pp_type t
     | Bgetopt t -> Format.fprintf fmt "getopt on %a" pp_type t
+    | Bfloor    -> pp_str fmt "floor"
+    | Bceil     -> pp_str fmt "ceil"
   in
 
   let pp_api_internal fmt = function
@@ -1155,6 +1159,15 @@ let pp_model fmt (model : model) =
       | Mgetopt x ->
         Format.fprintf fmt "getopt (%a)"
           f x
+
+      | Mfloor x ->
+        Format.fprintf fmt "floor (%a)"
+          f x
+
+      | Mceil x ->
+        Format.fprintf fmt "ceil (%a)"
+          f x
+
 
       (* crypto functions *)